rename macros.h to compile_date.h
